lib/model: Move test utilities to separate files
